requirements.txt
##pasgiref==3.3.4
Django==3.2.4
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
django-storages
